+++ src/extension.js
@@ -5,6 +5,7 @@ const { registerMacroAlias } = require("./macroAlias");
 const { registerOrphanKiller } = require("./orphanKiller");
 const { registerDiagnostics } = require("./diagnostics");
 const { registerToolDecorations } = require("./toolDecorations");
+const { registerChronometer } = require("./chronometer");
 
 function activate(context) {
 	registerFormatter(context);
@@ -14,6 +15,7 @@ function activate(context) {
 	registerOrphanKiller(context);
 	registerDiagnostics(context);
 	registerToolDecorations(context);
+	registerChronometer(context);
 }
 
 function deactivate() {}